requirements.txt
Flask
pandas
matplotlib
psycopg2-binary
python-dotenv
openai